refactor: extract shared cycle detection utility for swarm graph operations
