+++ assistant/src/swarm/orchestrator.ts
@@ -4,6 +4,7 @@ import type { SwarmWorkerBackend } from './worker-backend.js';
 import { runWorkerTask } from './worker-runner.js';
 import type { Provider } from '../providers/types.js';
 import { synthesizeResults } from './synthesizer.js';
+import { detectCycles } from './graph-utils.js';
 import { getLogger } from '../util/logger.js';
 
 const log = getLogger('swarm-orchestrator');
@@ -47,9 +48,9 @@ export async function executeSwarm(opts: ExecuteSwarmOptions): Promise<SwarmExec
   const startTime = Date.now();
 
   // Safety net: reject cyclic plans even if the caller skipped validation
-  const cyclePath = detectCycle(plan.tasks);
-  if (cyclePath) {
-    throw new Error(`Swarm plan contains a dependency cycle: ${cyclePath.join(' -> ')}`);
+  const cycleNodes = detectCycles(plan.tasks);
+  if (cycleNodes) {
+    throw new Error(`Swarm plan contains a dependency cycle: ${cycleNodes.join(' -> ')}`);
   }
 
   onStatus?.({ kind: 'plan_created', message: `Plan with ${plan.tasks.length} tasks` });
@@ -290,72 +291,6 @@ function blockDependents(
       blockDependents(depId, dependents, blocked, onStatus);
     }
   }
-}
-
-/**
- * DFS-based cycle detection. Returns the cycle path (e.g. ['a', 'b', 'c', 'a'])
- * if a cycle exists, or null if the graph is a valid DAG.
- */
-function detectCycle(tasks: SwarmTaskNode[]): string[] | null {
-  const adj = new Map<string, string[]>();
-  for (const t of tasks) {
-    adj.set(t.id, []);
-  }
-  for (const t of tasks) {
-    for (const dep of t.dependencies) {
-      // Edge from dep -> t.id (dep must finish before t)
-      adj.get(dep)?.push(t.id);
-    }
-  }
-
-  const WHITE = 0, GRAY = 1, BLACK = 2;
-  const color = new Map<string, number>();
-  for (const t of tasks) color.set(t.id, WHITE);
-
-  const parent = new Map<string, string>();
-
-  // Iterative DFS to avoid stack overflow on deep acyclic chains
-  for (const t of tasks) {
-    if (color.get(t.id) !== WHITE) continue;
-
-    const stack: Array<{ node: string; neighborIdx: number }> = [
-      { node: t.id, neighborIdx: 0 },
-    ];
-    color.set(t.id, GRAY);
-
-    while (stack.length > 0) {
-      const frame = stack[stack.length - 1];
-      const neighbors = adj.get(frame.node) ?? [];
-
-      if (frame.neighborIdx >= neighbors.length) {
-        // All neighbors visited — mark node as fully processed
-        color.set(frame.node, BLACK);
-        stack.pop();
-        continue;
-      }
-
-      const neighbor = neighbors[frame.neighborIdx];
-      frame.neighborIdx++;
-
-      if (color.get(neighbor) === GRAY) {
-        // Back edge found — reconstruct cycle path
-        const cycle = [neighbor];
-        for (let i = stack.length - 1; i >= 0; i--) {
-          cycle.push(stack[i].node);
-          if (stack[i].node === neighbor) break;
-        }
-        cycle.reverse();
-        return cycle;
-      }
-
-      if (color.get(neighbor) === WHITE) {
-        parent.set(neighbor, frame.node);
-        color.set(neighbor, GRAY);
-        stack.push({ node: neighbor, neighborIdx: 0 });
-      }
-    }
-  }
-  return null;
 }
 
 /** Resolves after `ms` milliseconds, or immediately if the signal fires first. */
